pyproject.toml
[project]
name = "readme-updater"
version = "0.1.0"
description = "My personal repository updater"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lxml>=6.0.2",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["toscrape"]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=7.0.0",
]
